requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
google-generativeai>=0.8.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0

# Web Framework
Flask>=3.0.0
Flask-Session>=0.5.0
Werkzeug>=3.0.0

# Utilities
langchain>=0.1.0
langgraph>=0.0.20
langsmith>=0.1.0

